Reorganize the headings when viewing journals
The resulting Markdown document now uses a level 1 heading for the date
range and a level 2 heading for each day. The headings in the journal
are incremented accordingly by 2. If only a single date is being viewed
only a single heading is added and the journal headings are incremented
by 1.